modules/payments-bridge/sear-payments-bridge-agents/src/main/java/com/silenteight/payments/bridge/agents/MatchtextFirstTokenOfAddressAgent.java
package com.silenteight.payments.bridge.agents;

import lombok.NonNull;

public class MatchtextFirstTokenOfAddressAgent {

  @NonNull
  public MatchtextFirstTokenOfAddressAgentResponse invoke(
      @NonNull MatchtextFirstTokenOfAddressAgentRequest
          matchtextFirstTokenOfAddressAgentRequest) {

    if (MatchtextFirstTokenOfAddressAgentHelper.isMatchtextFirstTokenOfAddress(
        matchtextFirstTokenOfAddressAgentRequest.getMatchingTexts(),
        matchtextFirstTokenOfAddressAgentRequest.getAddresses())) {
      return MatchtextFirstTokenOfAddressAgentResponse.YES;
    } else {
      return MatchtextFirstTokenOfAddressAgentResponse.NO;
    }
  }
}
